implemented Permissions from said Role + update session check

+++ controllers/delete_client.php
@@ -19,8 +19,8 @@ require_once('../models/init.php');
 if (session_status() == PHP_SESSION_NONE) {
     session_start();
 }
-if (!isset($_SESSION['username'])) {
-    header('Location:login.php');
+if (!isset($_SESSION['username']) || !isset($_SESSION['user_id']) || !isset($_SESSION['role'])) {
+    header('Location: ../views/login_view.php');
     exit();
 }
 
@@ -39,9 +39,18 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             exit();
         }
 
+        /*  *  *   *   PERMISSION    *    *   *   */
+        if ($_SESSION['role'] === 'client'){
+            echo json_encode(['message' => 'Vous n\'avez pas cette permission.']);
+            http_response_code(403);
+            exit;
+
+        }
         /*  *  *   *   ACTION    *    *   *   */
         $userDAL = new UserDAL();
+
         $result = $userDAL->delete($id);
+
 
         if ($result) {
